+++ src/app/open-api-generator/tests/generator.test.ts
@@ -84,6 +84,12 @@ const tests: {
     goldenFile: "./golden-files/adobe",
     baseUrl: "http://localhost:13191",
   },
+  {
+    name: "GenerateCode_Kubernetes",
+    oasFile: "./oas-docs/kubernetes.json",
+    goldenFile: "./golden-files/kubernetes",
+    baseUrl: "http://localhost:9191",
+  },
 
   // {
   //   name: "GenerateCode_LaunchDarklyApi",