+++ src/app/index.ts
@@ -9,11 +9,15 @@ import * as types from "./types";
 export async function runApp(args: types.GenerateCodeInput) {
   if (!context.getInstance().isOverwriteFilesEnabled()) {
     if (fs.existsSync(context.getInstance().getApiFilePath())) {
-      logger.fatal(`Error: overwriting is disbaled and api.ts file already exists at ${context.getInstance().getApiFilePath()}. To enable file overwrite, please set 'NDC_OAS_FILE_OVERWRITE' environment variable to true, or use the --overwrite flag`);
+      logger.fatal(
+        `Error: overwriting is disbaled and api.ts file already exists at ${context.getInstance().getApiFilePath()}. To enable file overwrite, please set 'NDC_OAS_FILE_OVERWRITE' environment variable to true, or use the --overwrite flag`,
+      );
       exit(0);
     }
     if (fs.existsSync(context.getInstance().getFunctionsFilePath())) {
-      logger.fatal(`Error: overwriting is disbaled and functions.ts file already exists at ${context.getInstance().getFunctionsFilePath()}. To enable file overwrite, please set 'NDC_OAS_FILE_OVERWRITE' environment variable to true, or use the --overwrite flag`);
+      logger.fatal(
+        `Error: overwriting is disbaled and functions.ts file already exists at ${context.getInstance().getFunctionsFilePath()}. To enable file overwrite, please set 'NDC_OAS_FILE_OVERWRITE' environment variable to true, or use the --overwrite flag`,
+      );
       exit(0);
     }
   }