+++ src/app/types.ts
@@ -5,5 +5,7 @@ export type GeneratedCode = {
 };
 
 export type GenerateCodeInput = {
-  openApiUri: string, headers?: string, baseUrl?: string,
-}
+  openApiUri: string;
+  headers?: string;
+  baseUrl?: string;
+};